fix: compilation errors - Byte/Int + duplicate companion
- WavHeaderWriter: writeByte(it.toInt()) for Byte→Int
- RecordingService: merge two companion objects into one

+++ app/src/main/java/com/kohei/summaryrecorder/recorder/WavHeaderWriter.kt
@@ -74,6 +74,6 @@ object WavHeaderWriter {
     }
 
     private fun RandomAccessFile.writeBytes(s: String) {
-        s.toByteArray(Charsets.US_ASCII).forEach { writeByte(it) }
+        s.toByteArray(Charsets.US_ASCII).forEach { writeByte(it.toInt()) }
     }
 }
